requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
cssselect==1.1.0
frozenlist==1.3.1
idna==3.4
lxml==4.9.1
multidict==6.0.2
PyYAML==6.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
yarl==1.8.1
